Add spinner in NButtonBase
It's not ideal from pretty-stand point, but works atm

https://github.com/Not-Endurance/not-timing-system/issues/356
