--- src/Not/Not.Blazor/Components/Mud/Buttons/Base/NButtonBase.cs
@@ -1,19 +1,0 @@
-﻿using MudBlazor;
-using Not.Localization;
-
-namespace Not.Blazor.Components;
-
-public abstract class NButtonBase : MudButton
-{
-    [Inject]
-    ILocalizer Localizer { get; set; } = default!;
-
-    [Parameter]
-    public string? Text { get; set; }
-
-    protected override void OnParametersSet()
-    {
-        ChildContent =
-            Text == null ? null : new RenderFragment(x => x.AddContent(0, Localizer.Get(Text)));
-    }
-}
